log gas throughput
