+++ crates/exex/exex/Cargo.toml
@@ -27,6 +27,7 @@ reth-payload-builder.workspace = true
 reth-evm.workspace = true
 reth-prune-types.workspace = true
 reth-revm.workspace = true
+reth-stages-api.workspace = true
 reth-stages-types.workspace = true
 
 ## async
