+++ crates/stages/api/src/metrics/mod.rs
@@ -1,5 +1,7 @@
+mod execution;
 mod listener;
 mod sync_metrics;
 
+pub use execution::format_gas_throughput;
 pub use listener::{MetricEvent, MetricEventsSender, MetricsListener};
 use sync_metrics::*;
